web-api/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
openai
pydantic
python-multipart